requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
django-bootstrap5==22.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0